Splitted spring configuration files
